Streamline search result component.
